Feat: Add email screen and form button widget
Added a new EmailScreen StatefulWidget that includes an AppBar, TextField for user input, and a FormButton. The state of the EmailScreen is managed with a TextEditingController which updates the username as it changes. Also added a new FormButton StatelessWidget that takes in a boolean to determine whether it's disabled or not. This button uses an AnimatedContainer and AnimatedDefaultTextStyle for smooth transitions when changing states.